smugmug_fs: gracefully degrade for video files with no mtime
Sometimes (as with the renames, I have no idea what triggers this)
SmugMug doesn't record a DateTimeModified for a video (presumably
other filetypes too, but we use MD5 checks for most other files) and
we would crash when checking the empty-string that we got
back. Instead just assume that the file isn't modified and move on.